ope-log-collector: Add fix for timeout error
Due to an issue in how the kubernetes_asyncio library handles timeouts,
the script would timeout every 5 minutes and we would have to reconnect.
This PR now uses a branch of kubernetes_asyncio that includes a fix
to this issue, and reformats the code so that it doesn't have to
reconnect the stream every 5 minutes.

Signed-off-by: Isaiah Stapleton <istaplet@redhat.com>